turismo_real_nuevo/requirements.txt

django == 2.2.3
pillow 
django-admin-interface
django-crispy-forms
django-rest-framework
pytest-django
django-pwa